--- Versione1/ServerImp.py
@@ -1,9 +1,0 @@
-import AbstractServer
-
-class server(AbstractServer):
-    def sendResponse(self,response):
-        return response
-
-if __name__ == "__main__":
-    s=server()
-
