+++ readme.md
@@ -1,10 +1,21 @@
 # Progetto Network Security
 
-Hi! I'm your first Markdown file in **StackEdit**.
-
 ## Versione0
 Prima versione del progetto 1:1:1 -> Un sensore,un attuatore in un campo.
 
 ## Version1 
 Seconda versione del progetto.
 Un campo è formato da più sotto-campi. All'interno del quale vi saranno 1 Sensore ed un 1 Attuatore
+
+## Rolling Version
+
+- **1.1.2**
+Versione che utilizza **GIT** 
+Per operare su questa versione fai una branch e successivamente una PULL REQUEST per unire il tuo lavoro
+
+- è stato aggiunto il file config.json dal quale si può effettuare il setup del server e del comportamento degli attuatori.
+- Ogni cosa ha una sua cartella adesso così da avere più ordine, se serve qualcosa condiviso creare una cartella utils. 
+- Inserita classe astratta per Sensori ed attuatori.
+- eliminato Campo.py
+
+
