+++ src/sensore/Sensore.py
