+++ src/sensore/SensoreAiocoap.py
