+++ src/server/globalConstants.py
